p4i integration Integrate 106053 from 3 to main: Integrate 102106 from 3.2 to 3: Fix compilation issue reported for beta1 with high warning level.
#23149

[git-p4: depot-paths = "//depot/qt/main/": change = 106056]

+++ extensions/activeqt/shared/types.cpp
@@ -216,7 +216,7 @@ bool QVariantToVARIANT( const QVariant &var, VARIANT &arg, const char *type )
 	    qvar.cast( proptype );
     }
 
-    switch ( qvar.type() ) {
+    switch ( (int)qvar.type() ) {
     case QVariant::String:
     case QVariant::CString:
 	arg.vt = VT_BSTR;
